Semifinal/requirements.txt
django==1.9.5
django-randomslugfield==0.3.0